Title - Music Player
Made the Music Player
Which is going to be one of the most used pieces of logic in this game,
so that was nice.
Using it right now to workout how I want the title screen to go, since I
want to have that experience nice and polished.

So all in all
The New stuff:
These have redundancy that I will trim and move around later, some of
their logic can be moved into a more generic class, but otherwise.
MusicSource - Handles playing audio clips to a source over a sequence
synchronised to the metronome.
MusicTrack - Holds sources, and manages their volumes, as well over a
sequence.
MusicPlayer - Has tracks, manages them.