Title - Improvements
Improved the Music Player, now capable of allowing an audio source to
continue until it's end. Meaning it now spawns audio sources with the
desired track playing, then tells them to destroy at the end of their
track length.

Improvements to the logo.

Starting to move code over to the SequencerGradient, which is my
solution to the previously mentioned code repetition I am having in
everything that goes across a sequence based on the synchronisation.