Title - The Night
The Night is bland and full of grey, I am fixing this. Stars will go a
long way, and the palette has been made a bit more colourful for the
night time, might play up more blue light.

Fireworks?
Lights moving along the streets?
Lights in general?

Also added the small wave normal to the terrain texture, massively
improved the look to me, just gave it a bit of wavyness on top of the
highly angular low poly look.