Nero - Splash
Nero's splash animation is now properly done, from writing his name
backwards to writing his name from end to start, to finally properly
writing his name, working on some kind of splash screen thing, need more
dynamic stuff, but that is a matter of layering.

Next is to polish up a bit more of the splash screen then move on to
stuff that is actually playable.