Nero - Splash
This is much closer to what I ultimately imagine for splash screens.
I need to work on transitioning in UI like it is being painted, or
'splattered' somehow, this is a matter of me figuring it out, not a
matter of it being difficult per say.

Cut some code out of ContraSynchronism, now have EdgeDetectionSequencer
or something like that, self explanatory.

Did some more Grunge stuff, making some 'blury' stripes, these are
currently used as the particle effect looking blury stripes that fly
behind Nero at the last parts of their splash.

As well, a whole bunch of meta data has been changed, as per the usual.
(Fix the git ignore)