General
I gave the Blackboard a strings literal class, so I no longer have to
remember if the blackboard key for the global synchroniser is
"Synchroniser" or "synchroniser" as well, I altered all references to
the string literal "Synchroniser" replacing them with the correct
Literals.Strings.Blackboard.Synchroniser

I made Controller inherit from NetworkBehaviour
I made PlayerController inherit from Controller

I made a MathHelper, currently allows me to isolate any 2 axis of a
Vector3 or Vector4 as a Vector2 (like HLSL) except since this is C# I
can't just make them friends, and so it had to be done in a static class
with static new object returning calls, which makes it significantly
slower than if I could just expose and return a lower dimensionality
vector from within a bigger one (assuming the Unity vector3 does in fact
inherit from Vector2 (which I see no way it does not))

StateMoveBehaviour
This behaviour manages the Move animation of any controller so that when
they begin moving or stop moving their isMoving parameter is properly
updated, as well when they begin moving their animation is offset to a
synchroniser (this means my animations can be synchronised to the music,
and apparently quite easily too).

I need to move all 'controller' logic to Controller, and ideally make a
C# pure "Character" class that enables me to make simple alterations to
a character.

As per the usual a bunch of useless meta data has been changed, and my
ignore continues to ignore the request to ignore these files, after
several attempts at clearing the cache, and a few at modifying the
ignore (or at least 1).